requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
distro==1.8.0
distro-info===0.18ubuntu0.18.04.1
feedparser==5.2.1
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
Jetson.GPIO==2.0.17
joblib==1.2.0
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
numpy==1.21.6
oauth==1.0.1
oauthlib==2.0.6
onboard==1.4.1
opencv-python==4.7.0.72
packaging==23.1
PAM==0.4.2
pandas==0.22.0
Pillow==9.5.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-build==0.17.6
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.6.3
ubuntu-drivers-common==0.0.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
xkit==0.0.0
youtube_dl==2018.3.14
zope.interface==4.3.2
